requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
google-generativeai>=0.3.0
google-ai-generativelanguage>=0.6.16,<0.7.0
chromadb>=0.4.0
sentence-transformers>=2.2.2
requests>=2.26.0
streamlit>=1.22.0
pandas>=1.3.0
